+++ programs/magicblock/src/args.rs
@@ -1,0 +1,42 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HandlerArgs {
+    pub escrow_index: u8,
+    pub data: Vec<u8>,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub struct CallHandlerArgs {
+    pub args: HandlerArgs,
+    pub destination_program: u8, // index of the account
+    pub accounts: Vec<u8>,       // indices of account
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub enum CommitTypeArgs {
+    Standalone(Vec<u8>), // indices on accounts
+    WithHandler {
+        committed_accounts: Vec<u8>, // indices of accounts
+        call_handlers: Vec<CallHandlerArgs>,
+    },
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub enum UndelegateTypeArgs {
+    Standalone,
+    WithHandler { call_handlers: Vec<CallHandlerArgs> },
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub struct CommitAndUndelegateArgs {
+    pub commit_type: CommitTypeArgs,
+    pub undelegate_type: UndelegateTypeArgs,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub enum MagicActionArgs {
+    L1Action(Vec<CallHandlerArgs>),
+    Commit(CommitTypeArgs),
+    CommitAndUndelegate(CommitAndUndelegateArgs),
+}
