+++ programs/magicblock/src/magicblock_instruction.rs
@@ -1,102 +1,9 @@
 use std::collections::HashMap;
 
-use magicblock_core::magic_program::MAGIC_CONTEXT_PUBKEY;
-use num_derive::{FromPrimitive, ToPrimitive};
 use serde::{Deserialize, Serialize};
-use solana_sdk::{
-    account::Account,
-    decode_error::DecodeError,
-    hash::Hash,
-    instruction::{AccountMeta, Instruction},
-    pubkey::Pubkey,
-    signature::Keypair,
-    signer::Signer,
-    transaction::Transaction,
-};
-use thiserror::Error;
+use solana_sdk::{account::Account, pubkey::Pubkey};
 
-use crate::{
-    mutate_accounts::set_account_mod_data,
-    validator::{validator_authority, validator_authority_id},
-};
-
-#[derive(
-    Error, Debug, Serialize, Clone, PartialEq, Eq, FromPrimitive, ToPrimitive,
-)]
-pub enum MagicBlockProgramError {
-    #[error("need at least one account to modify")]
-    NoAccountsToModify,
-
-    #[error("number of accounts to modify needs to match number of account modifications")]
-    AccountsToModifyNotMatchingAccountModifications,
-
-    #[error("The account modification for the provided key is missing.")]
-    AccountModificationMissing,
-
-    #[error("first account needs to be MagicBlock authority")]
-    FirstAccountNeedsToBeMagicBlockAuthority,
-
-    #[error("MagicBlock authority needs to be owned by system program")]
-    MagicBlockAuthorityNeedsToBeOwnedBySystemProgram,
-
-    #[error("The account resolution for the provided key failed.")]
-    AccountDataResolutionFailed,
-
-    #[error("The account data for the provided key is missing both from in-memory and ledger storage.")]
-    AccountDataMissing,
-
-    #[error("The account data for the provided key is missing from in-memory and we are not replaying the ledger.")]
-    AccountDataMissingFromMemory,
-
-    #[error("Tried to persist data that could not be resolved.")]
-    AttemptedToPersistUnresolvedData,
-
-    #[error("Tried to persist data that was resolved from storage.")]
-    AttemptedToPersistDataFromStorage,
-
-    #[error("Encountered an error when persisting account modification data.")]
-    FailedToPersistAccountModData,
-}
-
-impl<T> DecodeError<T> for MagicBlockProgramError {
-    fn type_of() -> &'static str {
-        "MagicBlockProgramError"
-    }
-}
-
-#[derive(Default, Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct AccountModification {
-    pub pubkey: Pubkey,
-    pub lamports: Option<u64>,
-    pub owner: Option<Pubkey>,
-    pub executable: Option<bool>,
-    pub data: Option<Vec<u8>>,
-    pub rent_epoch: Option<u64>,
-}
-
-impl From<(&Pubkey, &Account)> for AccountModification {
-    fn from(
-        (account_pubkey, account): (&Pubkey, &Account),
-    ) -> AccountModification {
-        AccountModification {
-            pubkey: *account_pubkey,
-            lamports: Some(account.lamports),
-            owner: Some(account.owner),
-            executable: Some(account.executable),
-            data: Some(account.data.clone()),
-            rent_epoch: Some(account.rent_epoch),
-        }
-    }
-}
-
-#[derive(Default, Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub(crate) struct AccountModificationForInstruction {
-    pub lamports: Option<u64>,
-    pub owner: Option<Pubkey>,
-    pub executable: Option<bool>,
-    pub data_key: Option<u64>,
-    pub rent_epoch: Option<u64>,
-}
+use crate::args::MagicActionArgs;
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
 pub enum MagicBlockInstruction {
@@ -163,47 +70,6 @@ pub enum MagicBlockInstruction {
     ScheduleAction(MagicActionArgs),
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct HandlerArgs {
-    pub escrow_index: u8,
-    pub data: Vec<u8>,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct CallHandlerArgs {
-    pub args: HandlerArgs,
-    pub destination_program: u8, // index of the account
-    pub accounts: Vec<u8>,       // indecis of account
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub enum CommitTypeArgs {
-    Standalone(Vec<u8>), // indices on accounts
-    WithHandler {
-        committed_accounts: Vec<u8>, // indices of accounts
-        call_handlers: Vec<CallHandlerArgs>,
-    },
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub enum UndelegateTypeArgs {
-    Standalone,
-    WithHandler { call_handlers: Vec<CallHandlerArgs> },
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct CommitAndUndelegateArgs {
-    pub commit_type: CommitTypeArgs,
-    pub undelegate_type: UndelegateTypeArgs,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub enum MagicActionArgs {
-    L1Action(Vec<CallHandlerArgs>),
-    Commit(CommitTypeArgs),
-    CommitAndUndelegate(CommitAndUndelegateArgs),
-}
-
 // TODO: why that exists?
 #[allow(unused)]
 impl MagicBlockInstruction {
@@ -229,169 +95,36 @@ impl MagicBlockInstruction {
     }
 }
 
-// -----------------
-// ModifyAccounts
-// -----------------
-pub fn modify_accounts(
-    account_modifications: Vec<AccountModification>,
-    recent_blockhash: Hash,
-) -> Transaction {
-    let ix = modify_accounts_instruction(account_modifications);
-    into_transaction(&validator_authority(), ix, recent_blockhash)
+#[derive(Default, Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub struct AccountModification {
+    pub pubkey: Pubkey,
+    pub lamports: Option<u64>,
+    pub owner: Option<Pubkey>,
+    pub executable: Option<bool>,
+    pub data: Option<Vec<u8>>,
+    pub rent_epoch: Option<u64>,
 }
 
-pub fn modify_accounts_instruction(
-    account_modifications: Vec<AccountModification>,
-) -> Instruction {
-    let mut account_metas =
-        vec![AccountMeta::new(validator_authority_id(), true)];
-    let mut account_mods: HashMap<Pubkey, AccountModificationForInstruction> =
-        HashMap::new();
-    for account_modification in account_modifications {
-        account_metas
-            .push(AccountMeta::new(account_modification.pubkey, false));
-        let account_mod_for_instruction = AccountModificationForInstruction {
-            lamports: account_modification.lamports,
-            owner: account_modification.owner,
-            executable: account_modification.executable,
-            data_key: account_modification.data.map(set_account_mod_data),
-            rent_epoch: account_modification.rent_epoch,
-        };
-        account_mods
-            .insert(account_modification.pubkey, account_mod_for_instruction);
+impl From<(&Pubkey, &Account)> for AccountModification {
+    fn from(
+        (account_pubkey, account): (&Pubkey, &Account),
+    ) -> AccountModification {
+        AccountModification {
+            pubkey: *account_pubkey,
+            lamports: Some(account.lamports),
+            owner: Some(account.owner),
+            executable: Some(account.executable),
+            data: Some(account.data.clone()),
+            rent_epoch: Some(account.rent_epoch),
+        }
     }
-    Instruction::new_with_bincode(
-        crate::id(),
-        &MagicBlockInstruction::ModifyAccounts(account_mods),
-        account_metas,
-    )
 }
 
-// -----------------
-// Schedule Commit
-// -----------------
-pub fn schedule_commit(
-    payer: &Keypair,
-    pubkeys: Vec<Pubkey>,
-    recent_blockhash: Hash,
-) -> Transaction {
-    let ix = schedule_commit_instruction(&payer.pubkey(), pubkeys);
-    into_transaction(payer, ix, recent_blockhash)
-}
-
-pub(crate) fn schedule_commit_instruction(
-    payer: &Pubkey,
-    pdas: Vec<Pubkey>,
-) -> Instruction {
-    let mut account_metas = vec![
-        AccountMeta::new(*payer, true),
-        AccountMeta::new(MAGIC_CONTEXT_PUBKEY, false),
-    ];
-    for pubkey in &pdas {
-        account_metas.push(AccountMeta::new_readonly(*pubkey, true));
-    }
-    Instruction::new_with_bincode(
-        crate::id(),
-        &MagicBlockInstruction::ScheduleCommit,
-        account_metas,
-    )
-}
-
-// -----------------
-// Schedule Commit and Undelegate
-// -----------------
-pub fn schedule_commit_and_undelegate(
-    payer: &Keypair,
-    pubkeys: Vec<Pubkey>,
-    recent_blockhash: Hash,
-) -> Transaction {
-    let ix =
-        schedule_commit_and_undelegate_instruction(&payer.pubkey(), pubkeys);
-    into_transaction(payer, ix, recent_blockhash)
-}
-
-pub(crate) fn schedule_commit_and_undelegate_instruction(
-    payer: &Pubkey,
-    pdas: Vec<Pubkey>,
-) -> Instruction {
-    let mut account_metas = vec![
-        AccountMeta::new(*payer, true),
-        AccountMeta::new(MAGIC_CONTEXT_PUBKEY, false),
-    ];
-    for pubkey in &pdas {
-        account_metas.push(AccountMeta::new_readonly(*pubkey, true));
-    }
-    Instruction::new_with_bincode(
-        crate::id(),
-        &MagicBlockInstruction::ScheduleCommitAndUndelegate,
-        account_metas,
-    )
-}
-
-// -----------------
-// Accept Scheduled Commits
-// -----------------
-pub fn accept_scheduled_commits(recent_blockhash: Hash) -> Transaction {
-    let ix = accept_scheduled_commits_instruction();
-    into_transaction(&validator_authority(), ix, recent_blockhash)
-}
-
-pub(crate) fn accept_scheduled_commits_instruction() -> Instruction {
-    let account_metas = vec![
-        AccountMeta::new_readonly(validator_authority_id(), true),
-        AccountMeta::new(MAGIC_CONTEXT_PUBKEY, false),
-    ];
-    Instruction::new_with_bincode(
-        crate::id(),
-        &MagicBlockInstruction::AcceptScheduleCommits,
-        account_metas,
-    )
-}
-
-// -----------------
-// Scheduled Commit Sent
-// -----------------
-pub fn scheduled_commit_sent(
-    scheduled_commit_id: u64,
-    recent_blockhash: Hash,
-) -> Transaction {
-    let ix = scheduled_commit_sent_instruction(
-        &crate::id(),
-        &validator_authority_id(),
-        scheduled_commit_id,
-    );
-    into_transaction(&validator_authority(), ix, recent_blockhash)
-}
-
-pub(crate) fn scheduled_commit_sent_instruction(
-    magic_block_program: &Pubkey,
-    validator_authority: &Pubkey,
-    scheduled_commit_id: u64,
-) -> Instruction {
-    let account_metas = vec![
-        AccountMeta::new_readonly(*magic_block_program, false),
-        AccountMeta::new_readonly(*validator_authority, true),
-    ];
-    Instruction::new_with_bincode(
-        *magic_block_program,
-        &MagicBlockInstruction::ScheduledCommitSent(scheduled_commit_id),
-        account_metas,
-    )
-}
-
-// -----------------
-// Utils
-// -----------------
-pub(crate) fn into_transaction(
-    signer: &Keypair,
-    instruction: Instruction,
-    recent_blockhash: Hash,
-) -> Transaction {
-    let signers = &[&signer];
-    Transaction::new_signed_with_payer(
-        &[instruction],
-        Some(&signer.pubkey()),
-        signers,
-        recent_blockhash,
-    )
+#[derive(Default, Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]
+pub struct AccountModificationForInstruction {
+    pub lamports: Option<u64>,
+    pub owner: Option<Pubkey>,
+    pub executable: Option<bool>,
+    pub data_key: Option<u64>,
+    pub rent_epoch: Option<u64>,
 }
