+++ programs/magicblock/src/utils/mod.rs
@@ -4,6 +4,7 @@ pub mod account_actions;
 pub mod accounts;
 #[cfg(not(test))]
 pub(crate) mod instruction_context_frames;
+pub mod instruction_utils;
 
 // NOTE: there is no low level SDK currently that exposes the program address
 //       we hardcode it here to avoid either having to pull in the delegation program
